pulled out test fixtures into common place
